Bugs completed

+++ app/src/main/java/com/example/ascen/utils/InputFilterMinMax.java
@@ -5,7 +5,7 @@ import android.text.Spanned;
 
 public class InputFilterMinMax implements InputFilter {
 
-    private int min, max;
+    private double min, max;
 
     public InputFilterMinMax(int min, int max) {
         this.min = min;
@@ -13,21 +13,21 @@ public class InputFilterMinMax implements InputFilter {
     }
 
     public InputFilterMinMax(String min, String max) {
-        this.min = Integer.parseInt(min);
-        this.max = Integer.parseInt(max);
+        this.min = Double.parseDouble(min);
+        this.max = Double.parseDouble(max);
     }
 
     @Override
     public CharSequence filter(CharSequence source, int start, int end, Spanned dest, int dstart, int dend) {
         try {
-            int input = Integer.parseInt(dest.toString() + source.toString());
+            double input = Double.parseDouble(dest.toString() + source.toString());
             if (isInRange(min, max, input))
                 return null;
         } catch (NumberFormatException nfe) { }     
         return "";
     }
 
-    private boolean isInRange(int a, int b, int c) {
+    private boolean isInRange(double a, double b, double c) {
         return b > a ? c >= a && c <= b : c >= b && c <= a;
     }
 }
